+++ src/index.js
@@ -14,14 +14,14 @@ const Statuses = {
 
 /**
  * Run migrations/tasks against database, specified by config.
- * @param {Object} config
+ * @param {Object} configuration
  * @param {Object[]} tasks
  * @param {string} tasks[].name - unique name of the task
  * @param {function} tasks[].operation - function, returning yieldable value (https://github.com/tj/co#yieldables)
  * @returns {Promise} resolved with hash (taskName: Status), or rejected with en error occurred
  */
-async function runMigrations(config, tasks) {
-    if (!isConfigurationValid(config)) {
+async function runMigrations(configuration, tasks) {
+    if (!isConfigurationValid(configuration)) {
         throw new IllegalConfigurationError();
     }
 
@@ -31,8 +31,8 @@ async function runMigrations(config, tasks) {
 
     const filteredTasks = filterUndefinedOrNullTasks(tasks);
 
-    const mongoClient = await MongoClient.connect(config.mongoUrl, config.mongoConnectionConfig);
-    const changelogCollection = await mongoClient.db(config.databaseName).createCollection('databasechangelog');
+    const mongoClient = await MongoClient.connect(configuration.mongoUrl, configuration.mongoConnectionConfig);
+    const changelogCollection = await mongoClient.db(configuration.databaseName).createCollection('databasechangelog');
     await changelogCollection.createIndex({name: 1}, {unique: true});
 
     const result = {};
@@ -74,7 +74,7 @@ async function processTask(task, changelogCollection) {
             dateExecuted: new Date(),
             md5sum: md5sum
         };
-        await changelogCollection.insert(appliedChange);
+        await changelogCollection.insertOne(appliedChange);
         status = Statuses.SUCCESSFULLY_APPLIED;
     }
     return status;